// discovery-provider/plugins/pedalboard/apps/postgraphile-trpc/src/prisma/schemas/objects/playlist_routesCountOrderByAggregateInput.schema.ts
import { z } from 'zod';
import { SortOrderSchema } from '../enums/SortOrder.schema';

import type { Prisma } from '@prisma/client';

const Schema: z.ZodType<Prisma.playlist_routesCountOrderByAggregateInput> = z
  .object({
    slug: z.lazy(() => SortOrderSchema).optional(),
    title_slug: z.lazy(() => SortOrderSchema).optional(),
    collision_id: z.lazy(() => SortOrderSchema).optional(),
    owner_id: z.lazy(() => SortOrderSchema).optional(),
    playlist_id: z.lazy(() => SortOrderSchema).optional(),
    is_current: z.lazy(() => SortOrderSchema).optional(),
    blockhash: z.lazy(() => SortOrderSchema).optional(),
    blocknumber: z.lazy(() => SortOrderSchema).optional(),
    txhash: z.lazy(() => SortOrderSchema).optional(),
  })
  .strict();

export const playlist_routesCountOrderByAggregateInputObjectSchema = Schema;